Move more (server-)code into logic module
A next step to extract to be less akka-dependent.
